+++ src/filters.rs
@@ -158,7 +158,7 @@ pub struct SimpleFilterParams {
 }
 
 impl SimpleFilterParams {
-    pub fn new(sample_rate: f32, gain_db: f32) -> SimpleFilterParams {
+    pub fn new(sample_rate: f64, gain_db: f32) -> SimpleFilterParams {
         SimpleFilterParams {
             sample_rate: sample_rate as FCoef,
             gain_db: gain_db as FCoef,
@@ -488,7 +488,7 @@ pub struct CrossfeedFilter {
 }
 
 impl CrossfeedFilter {
-    pub fn new(sample_rate: f32) -> CrossfeedFilter {
+    pub fn new(sample_rate: f64) -> CrossfeedFilter {
         CrossfeedFilter {
             level: 0.0,
             delay_ms: 0.3,
